Support for all possible subsets as input/output sets
Motivation: Currently, as the framework does not support any form of IOSet modeling,
we assume that all possible subsets of a modeled association could be a valid
IOSet specification! We don't want to implicitly constrain that in our code.

Thereby, also documenting a lot of previous assumptions and how we dealt with them.