Implement first version of control-flow-enablement
For this, currently only direct sequence flows between activities are considered
(no gateways are matched, which means we might be to "loose" compared to the real
control-flow-semantics).

Those sequence flows are then stored as "predecessors" for activities.
If the engine is then asked if an activity is (control-flow)-enabled, it will
lookup the execution history if there is a match for every predecessor
(same name and with an overlap of objects between the output of the historic
activity and at least one input combination of the current activity).