Support gateways and complex paths for finding predecessors (control-flow-enablement)
For finding predecessors of an activity, instead of only focussing on direct
sequence flows between activities, now all paths ending in the activity are
iteratively followed until they end (completely) or in predecessor activities.
Thereby, exclusive and parallel gateways are also evaluated accordingly.

For checking the control-flow-enablement the engine then differentiates between
exclusive and parallel behaviour of the predecessors.

Note that the "data overlap condition" (output list of predecessor needs at least
one object overlap with the input set of the current activity) still holds!
So just connecting two activities with a sequence flow without also connecting
data objects won't result in control flow enablement, as otherwise the engine
can't be sure if the activities are really part of the same "fragment instance"
(a concrete execution path of a fragment).