Support predecessors that write no data objects (control-flow-enablement)
Previously, there was the "data overlap condition" (output list of predecessor
needs at least one object overlap with the input set of the current activity).
So just connecting two activities with a sequence flow without also connecting
data objects won't result in control flow enablement, as otherwise the engine
can't be sure if the activities are really part of the same "fragment instance"
(a concrete execution path of a fragment).

Now the condition is extended to not only look at the output list of a predecessor
but also consider the input list. That is handy, if the predecessor reads the
same (remember overlap of at least one object) objects as the current activity
but does not write that object (not included in output).

But that is also enough for detecting that the predecessor and the current activity
are part of the same "fragment instance" (a concrete execution path of a fragment).