Fix too many usable objects returned for executing an activity
When an activity is not always control-flow-enabled, there was the bug that the
engine returned too many usable objects for executing it as soon as the control-
flow-enablement was fulfilled. This was caused by naively returning all suitable
objects (by state) but not checking whether these comply with control-flow-
enablement. This should be fixed now. =)